+++ src/components/ingredient-details/ingredient-details.module.css
@@ -1,39 +1,5 @@
-.modal {
-    display: flex;
-    position: relative;
-    flex-direction: column;
-    align-items: center;
+.container {
+    width: 100%;
+    max-width: 520px;
+    margin: 0 auto;
 }
-
-.title {
-    padding-top: 40px;
-    padding-left: 40px;
-    height: 64px;
-    display: flex;
-}
-
-.text{
-    justify-content: left;
-    margin:  auto auto auto 0;
-}
-
-.image {
-    height: 240px;
-    width: 480px;
-}
-
-.ingredients{
-    display: flex;
-    flex-direction: row;
-    padding-bottom: 60px;
-}
-
-.ingredient {
-    margin-right: 20px;
-    justify-content: center;
-}
-.ingredient:last-child{
-    margin-right: 0;
-}
-
-
